refactor(web): move api keys ui to developer
