+++ apps/web/src/app/(app)/developer/components/api-keys-section.tsx
@@ -1,3 +1,1 @@
-export function ApiKeysSection() {
-  return <div data-mount-point="api-keys-section" />;
-}
+export { ApiKeysSection } from "./api-keys";
