+++ apps/web/src/app/(app)/developer/components/api-keys/hooks/__tests__/use-api-keys.test.ts
@@ -1,7 +1,7 @@
 import { act, renderHook, waitFor } from "@testing-library/react";
 import { toast } from "sonner";
 import { beforeEach, describe, expect, it, vi } from "vitest";
-import { useApiKeys } from "@/app/connections/components/api-keys/hooks/use-api-keys";
+import { useApiKeys } from "@/app/developer/components/api-keys/hooks/use-api-keys";
 
 const listMock = vi.fn();
 const createMock = vi.fn();
