+++ apps/web/src/app/(app)/developer/components/api-keys/hooks/use-dialog-state.ts
@@ -5,8 +5,8 @@ import { useCallback, useEffect, useRef, useState } from "react";
 import type {
   ApiKeyRecord,
   DialogState,
-} from "@/app/connections/components/api-keys/types";
-import { DIALOG_CLEANUP_TIMEOUT } from "@/app/connections/components/api-keys/utils";
+} from "@/app/developer/components/api-keys/types";
+import { DIALOG_CLEANUP_TIMEOUT } from "@/app/developer/components/api-keys/utils";
 
 export function useDialogState(): DialogState {
   const [createDialogOpen, setCreateDialogOpen] = useState(false);
